+++ lib/install.bash
@@ -406,7 +406,7 @@ function install_python_tools {
 
 # Install extra javascript tools
 function install_js_tools {
-    exec_npm install -g eslint phantomjs-prebuilt;
+    exec_npm install -g eslint phantomjs-prebuilt stylelint stylelint-config-standard;
 }
 
 # install_python_prerequirements